+++ src/components/common/modal/ModalContext.tsx
@@ -1,0 +1,10 @@
+'use client';
+import { createContext } from 'react';
+
+type ModalContextType = {
+  closeModal: () => void;
+};
+
+const ModalContext = createContext<ModalContextType | null>(null);
+
+export default ModalContext;
